Generalize clipboard management
Remove the commands loadClip, showClipboard, clearClipboard;
and change saveClip (to be renamed) to modify a specific rewrites table
that represents a clipboard.

This design is strictly more general than before (show and clear can be
performed using the existing rewrites UI, and the load utterance can be
user-defined in a rewrites table). Also, it removes the separate
notion of "clipboard" which confusingly overlapped with rewrites.